ursina-chess/requirements.txt
ursina>=7.0.0
python-chess>=1.10.0
requests>=2.28.0
